confirm 组件的函数调用

+++ src/views/layout/components/header/header-search/history.vue
@@ -4,6 +4,7 @@ const EMITS_ITEM_CLICK = 'itemClick'
 <script setup>
 import {} from 'vue'
 import { useStore } from 'vuex'
+import { confirm } from '@/libs'
 
 const emits = defineEmits([EMITS_ITEM_CLICK])
 const store = useStore()
@@ -12,7 +13,9 @@ const store = useStore()
  * 删除所有记录
  */
 const onDeleteAllClick = () => {
-  store.commit('search/deleteAllHistory')
+  confirm('要删除所有历史记录吗？').then(() => {
+    store.commit('search/deleteAllHistory')
+  })
 }
 
 /**
